Organized the services package
